requirements.txt
Flask
flask-session
flask-cors
python-dotenv
gunicorn
requests
sentence-transformers
SQLAlchemy
msal
openai
langchain
langchain-community
langchain-openai
faiss-cpu
PyMuPDF
docx2txt
unstructured
pydantic
pytesseract
pdf2image
easyocr